TimeDependentDirichletBC employed in MovingTimeDependentProblem
geometry.h and MovingTimeDependentProblem.cpp contain only plotting options changes
In the main, TimeDependentDirichletBC seems to work properly only if the
"functor construction" of the TimeDependentExpression yielding the BC is used,
and not if the "inheriting construction" is used.
Future commits will have to fix.

firstDrop simulation comes from here.